Implement: Update remaining producer function signatures - specialized identifier types [#255]
Updated producer functions to use specialized identifier types (world.Id, channel.Id, _map.Id) instead of casting to primitive types. This improves type safety and consistency across the messaging layer.

Changes:
- Updated buddylist/producer.go - Remove byte() casts for WorldId
- Updated character/buff/producer.go - Use specialized types directly
- Updated drop/producer.go - Use WorldId, ChannelId, MapId without casts
- Updated cashshop/producer.go - Use specialized types in StatusEvent
- Updated chair/producer.go - Use specialized types in Command structs
- Updated character/expression/producer.go - Use specialized types
- Fixed cast issues in socket writers and consumers for I/O operations
- Fixed map/rest.go Extract function to return uint32 for character IDs

+++ atlas.com/channel/cashshop/producer.go
@@ -11,12 +11,12 @@ import (
 func CharacterEnterCashShopStatusEventProvider(actorId uint32, m _map.Model) model.Provider[[]kafka.Message] {
 	key := producer.CreateKey(int(actorId))
 	value := &cashshop.StatusEvent[cashshop.CharacterMovementBody]{
-		WorldId: byte(m.WorldId()),
+		WorldId: m.WorldId(),
 		Type:    cashshop.EventCashShopStatusTypeCharacterEnter,
 		Body: cashshop.CharacterMovementBody{
 			CharacterId: actorId,
-			ChannelId:   byte(m.ChannelId()),
-			MapId:       uint32(m.MapId()),
+			ChannelId:   m.ChannelId(),
+			MapId:       m.MapId(),
 		},
 	}
 	return producer.SingleMessageProvider(key, value)
@@ -25,12 +25,12 @@ func CharacterEnterCashShopStatusEventProvider(actorId uint32, m _map.Model) mod
 func CharacterExitCashShopStatusEventProvider(actorId uint32, m _map.Model) model.Provider[[]kafka.Message] {
 	key := producer.CreateKey(int(actorId))
 	value := &cashshop.StatusEvent[cashshop.CharacterMovementBody]{
-		WorldId: byte(m.WorldId()),
+		WorldId: m.WorldId(),
 		Type:    cashshop.EventCashShopStatusTypeCharacterExit,
 		Body: cashshop.CharacterMovementBody{
 			CharacterId: actorId,
-			ChannelId:   byte(m.ChannelId()),
-			MapId:       uint32(m.MapId()),
+			ChannelId:   m.ChannelId(),
+			MapId:       m.MapId(),
 		},
 	}
 	return producer.SingleMessageProvider(key, value)
